fix: compilation errors — float-to-int cast and missing coroutine imports
- TerminalView.java: explicit (int) cast for mFontWidth/mFontLineSpacing
  in cursor blink invalidation (float→int lossy conversion)
- TerminalService.kt: add coroutine imports (CoroutineScope, Dispatchers,
  launch) — were used via FQN but launch was unresolved

+++ core/terminal-view/src/main/java/com/termux/view/TerminalView.java
@@ -1507,11 +1507,11 @@ public final class TerminalView extends View {
                     mCursorVisible = !mCursorVisible;
                     em.setCursorBlinkState(mCursorVisible);
                     // Only invalidate the cursor cell region, not the entire screen.
-                    int cursorX = em.getCursorCol() * mRenderer.mFontWidth;
-                    int cursorY = (em.getCursorRow() + mTopRow + em.getScreen().getActiveRows() - em.mRows)
-                                  * mRenderer.mFontLineSpacing;
+                    int cursorX = (int) (em.getCursorCol() * mRenderer.mFontWidth);
+                    int cursorY = (int) ((em.getCursorRow() + mTopRow + em.getScreen().getActiveRows() - em.mRows)
+                                  * mRenderer.mFontLineSpacing);
                     invalidate(cursorX, cursorY,
-                        cursorX + mRenderer.mFontWidth, cursorY + mRenderer.mFontLineSpacing);
+                        cursorX + (int) mRenderer.mFontWidth, cursorY + (int) mRenderer.mFontLineSpacing);
                 }
             } finally {
                 // Recall the Runnable after mBlinkRate milliseconds to toggle the blink state
